refactor: rename to loadingButton and pull into folder
