+++ src/components/Button/LoadingButton/Spinner.tsx
@@ -1,0 +1,20 @@
+import styled from "styled-components";
+
+export const Spinner = styled.span`
+    width: 21px;
+    height: 21px;
+    border: 3px solid #FFF;
+    border-bottom-color: transparent;
+    border-radius: 50%;
+    display: inline-block;
+    box-sizing: border-box;
+    animation: rotation 1s linear infinite;
+    }
+    @keyframes rotation {
+    0% {
+        transform: rotate(0deg);
+    }
+    100% {
+        transform: rotate(360deg);
+    }
+`;
